+++ src/components/SudokuBox.jsx
@@ -8,7 +8,7 @@ import SudokuNumberBox from "./SudokuNumberBox";
 
 const useStyles = makeStyles((theme) => {
     const light = theme.palette.type === "light";
-    const mainbordercolor = light ? theme.palette.info.main : theme.palette.primary.dark;
+    const mainbordercolor = light ? theme.palette.info.main : theme.palette.primary.light;
     return {
         subbox: { textAlign: "center", height: "100%" },
         grid: { borderRight: theme.spacing(0.125) + "px solid " + mainbordercolor },
